+++ BackEnd/EG.Domain/DTOs/Requests/General/MenuRoleDto.cs
@@ -6,7 +6,7 @@ namespace EG.Domain.DTOs.Requests.General
 {
     public class MenuRoleDto
     {
-        public long FkidMenuSis { get; set; }
+        public int FkidMenuSis { get; set; }
 
         public string RoleId { get; set; }
